fix controller structure for better representation
