+++ src/main/java/com/kulakyokedici/kulakliksitesi/controller/AdminUserController.java
@@ -1,63 +1,32 @@
 package com.kulakyokedici.kulakliksitesi.controller;
 
-import java.util.List;
-import java.util.Set;
-
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
-import com.kulakyokedici.kulakliksitesi.objects.data.dto.request.SellerCreateRequest;
-import com.kulakyokedici.kulakliksitesi.objects.data.dto.request.SellerUpdateRequest;
-import com.kulakyokedici.kulakliksitesi.objects.data.dto.request.ShopperCreateRequest;
-import com.kulakyokedici.kulakliksitesi.objects.data.dto.request.ShopperDetailsUpdateRequest;
-import com.kulakyokedici.kulakliksitesi.objects.data.dto.request.ShopperUpdateRequest;
-import com.kulakyokedici.kulakliksitesi.objects.data.dto.request.UserCreateRequest;
 import com.kulakyokedici.kulakliksitesi.objects.data.dto.request.UserUpdateRequest;
-import com.kulakyokedici.kulakliksitesi.objects.data.dto.response.AdminResponse;
-import com.kulakyokedici.kulakliksitesi.objects.data.dto.response.ItemSummaryResponse;
-import com.kulakyokedici.kulakliksitesi.objects.data.dto.response.SellerDetailedResponse;
-import com.kulakyokedici.kulakliksitesi.objects.data.dto.response.SellerResponse;
-import com.kulakyokedici.kulakliksitesi.objects.data.dto.response.ShopperResponse;
 import com.kulakyokedici.kulakliksitesi.objects.data.dto.response.UserResponse;
-import com.kulakyokedici.kulakliksitesi.objects.exception.InsufficientParametersException;
-import com.kulakyokedici.kulakliksitesi.service.AdminService;
-import com.kulakyokedici.kulakliksitesi.service.ItemService;
-import com.kulakyokedici.kulakliksitesi.service.SellerService;
-import com.kulakyokedici.kulakliksitesi.service.ShopperService;
 import com.kulakyokedici.kulakliksitesi.service.UserService;
 
 import jakarta.validation.Valid;
 
 @RestController
 @RequestMapping("api/admin")
-public class AdminController
+public class AdminUserController
 {
 	private UserService userService;
-	private ShopperService shopperService;
-	private SellerService sellerService;
-	private ItemService itemService;
-	private AdminService adminService;
 	
 	@Autowired
-	public AdminController(UserService userService,
-			ShopperService shopperService,
-			SellerService sellerService,
-			ItemService itemService,
-			AdminService adminService)
+	public AdminUserController(
+			UserService userService)
 	{
 		this.userService = userService;
-		this.shopperService = shopperService;
-		this.sellerService = sellerService;
-		this.itemService = itemService;
-		this.adminService = adminService;
 	}
 	
 	/*
@@ -81,116 +50,17 @@ public class AdminController
 		return ResponseEntity.ok(userResponse);
 	}
 	
-	@GetMapping("/admins")
-	public ResponseEntity<?> getAdmin(
-			@RequestParam(name = "username", required = false) String username,
-			@RequestParam(name = "email", required = false) String email)
-	{
-		AdminResponse adminResponse = null;
-		
-		if(username != null && !username.isBlank())
-			adminResponse = adminService.getByUsername(username);
-		else if(email != null && !email.isBlank())
-			adminResponse = adminService.getByEmail(email);
-		else
-			return ResponseEntity.ok(adminService.getAll());
-		
-		return ResponseEntity.ok(adminResponse);
-	}
-	
-	@GetMapping("/sellers")
-	public ResponseEntity<?> getSeller(
-			@RequestParam(name = "username", required = false) String username,
-			@RequestParam(name = "email", required = false) String email,
-			@RequestParam(name = "company_name", required = false) String companyName)
-	{
-		SellerResponse sellerResponse = null;
-		
-		if(username != null && !username.isBlank())
-			sellerResponse = sellerService.getByUsername(username);
-		else if(email != null && !email.isBlank())
-			sellerResponse = sellerService.getByEmail(email);
-		else if(companyName != null && !companyName.isBlank())
-			sellerResponse = sellerService.getByCompanyName(companyName);
-		else
-			return ResponseEntity.ok(sellerService.getAll());
-		
-		return ResponseEntity.ok(sellerResponse);
-	}
-	
-	// isim soyisim araması eklenecek
-	@GetMapping("/shoppers")
-	public ResponseEntity<?> getShopper(
-			@RequestParam(name = "username", required = false) String username,
-			@RequestParam(name = "email", required = false) String email)
-	{
-		ShopperResponse shopperResponse = null;
-		
-		if(username != null && !username.isBlank())
-			shopperResponse = shopperService.getByUsername(username);
-		else if(email != null && !email.isBlank())
-			shopperResponse = shopperService.getByEmail(email);
-		else
-			return ResponseEntity.ok(shopperService.getAll());
-		
-		return ResponseEntity.ok(shopperResponse);
-	}
-	
 	@GetMapping("/users/{id}")
 	public ResponseEntity<UserResponse> getUserById(@PathVariable Long id)
 	{
 		return ResponseEntity.ok(userService.getById(id));
 	}
 	
-	@GetMapping("/shoppers/{id}")
-	public ResponseEntity<ShopperResponse> getShopperById(@PathVariable Long id)
-	{
-		return ResponseEntity.ok(shopperService.getById(id));
-	}
-	
-	@GetMapping("/sellers/{id}")
-	public ResponseEntity<SellerDetailedResponse> getSellerById(@PathVariable Long id)
-	{
-		return ResponseEntity.ok(sellerService.getDetailedResponseById(id));
-	}
-	
-	@GetMapping("/admins/{id}")
-	public ResponseEntity<AdminResponse> getAdminById(@PathVariable Long id)
-	{
-		return ResponseEntity.ok(adminService.getById(id));
-	}
-	
-	@GetMapping("/sellers/{sellerId}/items")
-	public ResponseEntity<Set<ItemSummaryResponse>> getItemsBySellerId(@PathVariable Long sellerId)
-	{
-		return ResponseEntity.ok(itemService.getAllBySellerId(sellerId));
-	}
 	
 	/*
 	 * POST istekleri
 	 * yeni kaynak eklemek için.
 	 */
-	
-	@PostMapping("/admins")
-	public ResponseEntity<Void> addAdmin(@Valid @RequestBody UserCreateRequest newAdmin)
-	{
-		adminService.add(newAdmin);
-		return ResponseEntity.ok().build();
-	}
-	
-	@PostMapping("/sellers")
-	public ResponseEntity<Void> addSeller(@Valid @RequestBody SellerCreateRequest req)
-	{	
-		sellerService.add(req);
-		return ResponseEntity.ok().build();
-	}
-	
-	@PostMapping("/shoppers")
-	public ResponseEntity<Void> addShopper(@Valid @RequestBody ShopperCreateRequest req)
-	{
-		shopperService.add(req);
-		return ResponseEntity.ok().build();
-	}
 	
 	/*
 	 * PUT istekleri
@@ -202,38 +72,6 @@ public class AdminController
 			@Valid @RequestBody UserUpdateRequest newUser)
 	{
 		userService.update(userId, newUser);
-		return ResponseEntity.noContent().build();
-	}
-	
-	@PutMapping("/sellers/{sellerId}")
-	public ResponseEntity<Void> updateSeller(@PathVariable Long sellerId,
-			@Valid @RequestBody SellerUpdateRequest newSeller)
-	{
-		sellerService.update(sellerId, newSeller);
-		return ResponseEntity.noContent().build();
-	}
-	
-	@PutMapping("/shoppers/{shopperId}")
-	public ResponseEntity<Void> updateShopper(@PathVariable Long shopperId, 
-			@Valid @RequestBody ShopperUpdateRequest newShopper)
-	{
-		shopperService.update(shopperId, newShopper);
-		return ResponseEntity.noContent().build();
-	}
-	
-	@PutMapping("/shoppers/{shopperId}/details")
-	public ResponseEntity<Void> updateShopperDetails(@PathVariable Long shopperId, 
-			@Valid @RequestBody ShopperDetailsUpdateRequest newShopperDetails)
-	{
-		shopperService.updateDetails(shopperId, newShopperDetails);
-		return ResponseEntity.noContent().build();
-	}
-	
-	@PutMapping("/admins/{adminId}")
-	public ResponseEntity<Void> updateAdmin(@PathVariable Long adminId,
-			@Valid @RequestBody UserUpdateRequest newAdmin)
-	{
-		adminService.update(adminId, newAdmin);
 		return ResponseEntity.noContent().build();
 	}
 }
